+++ modules/settings.js
@@ -1,4 +1,5 @@
 import * as MAESTRO from "./config.js";
+import { _checkForCriticalPlaylist, _checkForFailurePlaylist } from "./misc.js";
 
 export const registerModuleSettings = function() {
 
@@ -112,8 +113,69 @@ export const registerModuleSettings = function() {
         hint: "SETTINGS.DisableDiceSoundH",
         scope: "world",
         type: Boolean,
-        default: "",
+        default: false,
         config: true,
+        onChange: s => {
+
+        }
+    }),
+
+    game.settings.register(MAESTRO.MODULE_NAME, MAESTRO.SETTINGS_KEYS.Misc.enableCriticalSuccessFailureTracks, {
+        name: "SETTINGS.EnableCriticalSuccessFailureTracksN",
+        hint: "SETTINGS.EnableCriticalSuccessFailureTracksH",
+        scope: "world",
+        type: Boolean,
+        default: false,
+        config: true,
+        onChange: s => {
+
+        }
+    }),
+
+    game.settings.register(MAESTRO.MODULE_NAME, MAESTRO.SETTINGS_KEYS.Misc.createCriticalSuccessPlaylist, {
+        name: "SETTINGS.CreateCriticalPlaylistN",
+        hint: "SETTINGS.CreateCriticalPlaylistH",
+        scope: "world",
+        type: Boolean,
+        default: false,
+        config: true,
+        onChange: s => {
+            if (!game.settings.get(MAESTRO.MODULE_NAME, MAESTRO.SETTINGS_KEYS.Misc.enableCriticalSuccessFailureTracks)) {
+                return;
+            }
+
+            _checkForCriticalPlaylist();
+        }
+    }),
+
+    game.settings.register(MAESTRO.MODULE_NAME, MAESTRO.SETTINGS_KEYS.Misc.createCriticalFailurePlaylist, {
+        name: "SETTINGS.CreateFailurePlaylistN",
+        hint: "SETTINGS.CreateFailurePlaylistH",
+        scope: "world",
+        type: Boolean,
+        default: false,
+        config: true,
+        onChange: s => {
+            if (!game.settings.get(MAESTRO.MODULE_NAME, MAESTRO.SETTINGS_KEYS.Misc.enableCriticalSuccessFailureTracks)) {
+                return;
+            }
+
+            _checkForFailurePlaylist();
+        }
+    }),
+
+    game.settings.register(MAESTRO.MODULE_NAME, MAESTRO.SETTINGS_KEYS.Misc.criticalSuccessFailureTracks, {
+        name: "SETTINGS.CriticalSuccessFailureTracksN",
+        hint: "SETTINGS.CriticalSuccessFailureTracksH",
+        scope: "world",
+        type: Object,
+        default: {
+            criticalSuccessPlaylist: game.playlists ? game.playlists.entities.find(p => p.name === MAESTRO.DEFAULT_CONFIG.Misc.criticalSuccessPlaylistName) : "",
+            criticalSuccessSound: "",
+            criticalFailurePlaylist: game.playlists ? game.playlists.entities.find(p => p.name === MAESTRO.DEFAULT_CONFIG.Misc.criticalFailurePlaylistName) : "",
+            criticalFailureSound: ""
+        },
+        config: false,
         onChange: s => {
 
         }